Removing some commented out code plus variables that were declared but not used.
