+++ config/connection.js
@@ -16,18 +16,4 @@ const sequelize = process.env.JAWSDB_URL
 
 module.exports = sequelize;
 
-// const Sequelize = require('sequelize');
-// require('dotenv').config();
 
-// const sequelize = new Sequelize(
-//   process.env.DB_NAME,
-//   process.env.DB_USER,
-//   process.env.DB_PASSWORD,
-//   {
-//     host: 'localhost',
-//     dialect: 'mysql',
-//     port: 3306,
-//   }
-// );
-
-// module.exports = sequelize;
