+++ controllers/api/wishlistRoutes.js
@@ -1,5 +1,4 @@
 const router = require("express").Router();
-const e = require("express");
 const { Wishlist } = require("../../models");
 const { User } = require("../../models");
 
@@ -13,22 +12,6 @@ router.get("/", async (req, res) => {
   res.json(allWishlist);
 });
 
-
-// get single wishlist by id and return selected wishlist as JSON payload
-// router.get("/:id", async (req, res) => {
-//   try {
-//     const selectedWishlist = await Wishlist.findOne({
-//       where: { id: req.params.id },
-//     });
-//     if (!selectedWishlist) {
-//       res.status(404).json({ message: "Wishlist does not exist" });
-//       return;
-//     }
-//     res.status(200).json(selectedWishlist);
-//   } catch (err) {
-//     res.status(500).json(err);
-//   }
-// });
 
 // post a new wishlist item to the database
 // this is the api/wishlist/ endpoint
@@ -75,7 +58,7 @@ router.get('/:id', async (req, res) => {
 });
 
 
-// update a wishlist by id and return updated wishlist as JSON payload.  this is the /api/wishlist/:id endpoint
+// update a wishlist by id and return updated wishlist as JSON payload.  this is the /api/wishlist/:id route
 /* put should look like this...
       {
         "id": "6",
@@ -99,7 +82,7 @@ router.put("/:id", async (req, res) => {
   }
 });
 
-// delete a wishlist by its `id` value.  this is the /api/wishlist/:id endpoint
+// delete a wishlist by its `id` value.  this is the /api/wishlist/:id route
 router.delete("/:id", async (req, res) => {
   try {
     const [deletedRows] = await Wishlist.destroy({
@@ -118,9 +101,31 @@ router.delete("/:id", async (req, res) => {
 });
 
 
+// renders userWishlist.handlebars using friend's name.  this is the api/wishlist/friendFoundWishlist route
+router.get("/friendFoundWishlist/:friendEmail", async (req, res) => {
+  try {
+    const userData = await User.findOne({
+      where: { email: req.params.friendEmail },
+      include: [{ model: Wishlist }],
+    });
 
+    const user = userData.get({ plain: true });
 
+    if (!user) {
+      res.status(404).json({ message: "User does not exist" });
+      return;
+    }
 
+    res.render("userWishlist", { ...user });
+  } catch (err) {
+    // res.status(500).json(err);
+    res.status(500).json("Ouch");
+  }
+});
+
+router.get("/updateWishlist", (req, res) => {
+  res.render("wishlistUpdate");
+});
