+++ controllers/homeRoutes.js
@@ -1,5 +1,4 @@
 const router = require("express").Router();
-const session = require("express-session");
 const User = require("../models/User");
 const Wishlist = require("../models/Wishlist");
 const withAuth = require("../utils/auth");
@@ -40,12 +39,12 @@ router.get("/signup", (req, res) => {
   res.render("signup");
 });
 
-// renders addWishlistItem.handlebars.  this is the /post endpoint
+// renders addWishlistItem.handlebars.  this is the /post route
 router.get("/post", (req, res) => {
   res.render("addWishlistItem");
 });
 
-// renders userWishlist.handlebars using current session user ID.  this is the /wishlist endpoint
+// renders userWishlist.handlebars using current session user ID.  this is the /wishlist route
 router.get("/wishlist", async (req, res) => {
   try {
     const userData = await User.findByPk(req.session.user_id, {
@@ -64,35 +63,11 @@ router.get("/wishlist", async (req, res) => {
   }
 });
 
-// renders getFriendWishlist.handlebars.  Is used for selecting a friend's wishlist.  this is the /friendWishlist endpoint
+// renders getFriendWishlist.handlebars.  Is used for selecting a friend's wishlist.  this is the /friendWishlist route
 router.get("/friendWishlist", (req, res) => {
   res.render("getFriendWishlist");
 });
 
-// renders userWishlist.handlebars using friend's name.  this is the /friendFoundWishlist endpoint
-router.get("/friendFoundWishlist/:friendEmail", async (req, res) => {
-  try {
-    const userData = await User.findOne({
-      where: { email: req.params.friendEmail },
-      include: [{ model: Wishlist }],
-    });
 
-    const user = userData.get({ plain: true });
-
-    if (!user) {
-      res.status(404).json({ message: "User does not exist" });
-      return;
-    }
-
-    res.render("userWishlist", { ...user });
-  } catch (err) {
-    // res.status(500).json(err);
-    res.status(500).json("Ouch");
-  }
-});
-
-router.get("/updateWishlist", (req, res) => {
-  res.render("wishlistUpdate");
-});
 
 module.exports = router;
